Map and autogetting user IP functional
Added a feature for auto-filling the input with the user’s IP data upon consent (when the application is launched for the first time). Data is stored in local storage.

Added basic map functionality. Creation on zero coordinates. Ability to add a base marker and move the center to the desired coordinates. Works in conjunction with the current receipt of user IP data.