Map addon, theme switching functional, and small refactoring
Added scss and js theme changing functionality. Dark theme colors are not defined.

Added the ability to add a custom marker to the map when initializing the map. A function that connects the position of the view on the map and the addition of a marker (installation at the same coordinates).

Some functions have been moved to separate files in utils.